Combine ServicesScreen and ServiceEdit
The ServicesScreen was the only thing using ServiceEdit, and ServiceEdit is likely to be completely rewritten anyway
