--- frontend/src/components/service-edit.module.css
@@ -1,55 +1,0 @@
-#input-container {
-	display: flex;
-	flex-direction: column;
-	flex-grow: 1;
-	background: #464142aa;
-	margin: 3pt;
-	padding: 3pt;
-	border-radius: 5pt;
-	max-width: 220pt;
-	min-width: 110pt;
-}
-
-#buttons {
-	display: flex;
-	flex-wrap: nowrap;
-	gap: 5pt;
-	margin: 3pt;
-}
-
-.delete {
-	background: red;
-}
-
-.delete:hover {
-	background: #f99;
-}
-
-.submit,
-.delete {
-	color: black;
-	padding: 5pt;
-	font-size: 12pt;
-	border-radius: 6pt;
-	width: 100%;
-}
-
-.input {
-	padding: 5pt;
-	background: #0005;
-	color: white;
-	border: none;
-	border-radius: 5pt;
-	transition-duration: 0.5s;
-	flex: 1;
-	margin: 3pt;
-}
-
-.input:hover {
-	box-shadow: 0pt 0pt 3pt #555a;
-	transition-duration: 0s;
-}
-
-.input:focus {
-	outline: solid 2pt #aaa;
-}
